0.0.12.2.1版
添加了题型和Focus分析材料, 更改了问题读取的方式。

+++ src/com/question/Perceptron.java
@@ -87,7 +87,7 @@ public class Perceptron {
     }
 
     public double multipleWandX(int index) {
-        double ret = (double) 0.0;
+        double ret = 0.0;
         for (int i = 0; i < dim; i++)
             ret += w[i] * x[index][i];
         return ret;
